+++ mmdeploy/utils/__init__.py
@@ -7,6 +7,7 @@ from .config_utils import (cfg_apply_marks, get_backend, get_backend_config,
                            is_dynamic_batch, is_dynamic_shape, load_config)
 from .constants import Backend, Codebase, Task
 from .device import parse_cuda_device_id, parse_device_id
+from .utils import target_wrapper
 
 __all__ = [
     'is_dynamic_batch', 'is_dynamic_shape', 'get_task_type', 'get_codebase',
@@ -14,5 +15,6 @@ __all__ = [
     'get_onnx_config', 'get_partition_config', 'get_calib_config',
     'get_calib_filename', 'get_common_config', 'get_model_inputs',
     'cfg_apply_marks', 'get_input_shape', 'parse_device_id',
-    'parse_cuda_device_id', 'get_codebase_config', 'get_backend_config'
+    'parse_cuda_device_id', 'get_codebase_config', 'get_backend_config',
+    'target_wrapper'
 ]
