+++ mmdeploy/utils/utils.py
@@ -1,0 +1,23 @@
+import sys
+import traceback
+import logging
+
+
+def target_wrapper(target, log_level, ret_value, *args, **kwargs):
+    logger = logging.getLogger()
+    logging.basicConfig(
+        format='%(asctime)s,%(name)s %(levelname)-8s'
+        ' [%(filename)s:%(lineno)d] %(message)s',
+        datefmt='%Y-%m-%d:%H:%M:%S')
+    logger.level
+    logger.setLevel(log_level)
+    if ret_value is not None:
+        ret_value.value = -1
+    try:
+        result = target(*args, **kwargs)
+        if ret_value is not None:
+            ret_value.value = 0
+        return result
+    except Exception as e:
+        logging.error(e)
+        traceback.print_exc(file=sys.stdout)
